dgus: m300: schedule m300 at proper time
This makes some modifications to allow us to get the proper "print_time" to
start the M300 and then schedule the UART command at that same "print_time".
With this, the beep is played immediately when the gcode parser performs
lookahead.  Multiple beeps are played at the same time and G4 commands did
not perform waits between beeps.

We should still add support for scheduling longer beeps by refilling the
beep duration register when it gets low.  That can be done by using a
reactor timer.  That should be at a lower layer though so that all of the
code can take advantage of long beeps, not just M300.